ADD Gestion du paiement de commande

+++ src/main/java/com/pa/shoploc/bo/PaiementHisto.java
@@ -32,7 +32,7 @@ public class PaiementHisto {
     @JoinColumn(name="username")
     Client user;
 
-    float money;
+    double money;
 
 
 }
